+++ src/main/java/com/bnhp/stock/model/document/Portfolio.java
@@ -3,8 +3,6 @@ package com.bnhp.stock.model.document;
 import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigDecimal;
-
 @Data
 @Document(value = "portfolio")
 public class Portfolio {
@@ -13,7 +11,7 @@ public class Portfolio {
     private String userId;
     private String stockId;
     private Integer quantity;
-    private BigDecimal averagePrice;
-    private BigDecimal totalInvestment;
+    private Double averagePrice;
+    private Double totalInvestment;
 
 }
